refactor(reborn): move automation tests to checker-recognized tests modules
scripts/check_no_panics.py exempts test code by path (tests/ dir
component or tests.rs filename), not by #[cfg(test)] gating, so the
src/automation_resolver_tests.rs split tripped the no-panics CI job.
Restructure to the blessed layout: automation.rs's inline test mod
moves to src/automation/tests.rs and the resolver tests to
src/automation/tests/resolver_tests.rs. automation.rs drops to 337
lines.
