+++ crates/ironclaw_reborn_composition/src/automation/tests/resolver_tests.rs
